### Click package update: - GainAMP 2 click - Gaussmeter click - GeoMagnetic click - GMR Angle click - GNSS2 click - GNSS3 click - GNSS4 click - GNSS5 click - GNSS7 click - GNSS click - GNSS RTK click - GNSS ZOE click - G2C 3G click - G2C click - GPS 3 click - GPS 4 click - GPS click - Grid-Eye click - GSM 2 click - GSM 3 click - GSM 4 click - GSM click - GSM/GNSS click - GSM-GPS click - GSR click - Gyro 2 click - Gyro 3 click - Gyro 4 click - Gyro 5 click - Gyro click - Hall Current 2 click - Hall Current 3 click - Hall Current 4 click - Hall Current 5 click - Hall Current 6 click - Hall Current 8 25A click - Hall Current 8 120A click - Hall Current 9 click - Hall Current click - Hall Switch click - HAPTIC 2 click - HAPTIC click - H-Bridge 2 click - H-Bridge 3 click - H-Bridge 4 click - H-Bridge 5 click - H-Bridge 7 click

+++ clicks/gps/lib/src/gps.c
@@ -57,8 +57,8 @@ void gps_cfg_setup ( gps_cfg_t *cfg )
     
     // Additional gpio pins
 
-     cfg->reset = HAL_PIN_NC;
-     cfg->tmpls = HAL_PIN_NC;
+    cfg->reset = HAL_PIN_NC;
+    cfg->tmpls = HAL_PIN_NC;
 
     cfg->baud_rate      = 9600;
     cfg->data_bit       = UART_DATA_BITS_DEFAULT;
@@ -93,15 +93,13 @@ GPS_RETVAL gps_init ( gps_t *ctx, gps_cfg_t *cfg )
 
     // Output pins 
 
-     digital_out_init( &ctx->reset, cfg->reset );
+    digital_out_init( &ctx->reset, cfg->reset );
 
     // Input pins
    
-     digital_in_init( &ctx->tmpls, cfg->tmpls );
+    digital_in_init( &ctx->tmpls, cfg->tmpls );
     
-
     return GPS_OK;
-
 }
 
 void gps_module_wakeup ( gps_t *ctx )
@@ -130,7 +128,7 @@ void gps_generic_write ( gps_t *ctx, char *data_buf, uint16_t len )
     uart_write( &ctx->uart, data_buf, len );
 }
 
-int16_t gps_generic_read ( gps_t *ctx, char *data_buf, uint16_t max_len )
+int32_t gps_generic_read ( gps_t *ctx, char *data_buf, uint16_t max_len )
 {
     return uart_read( &ctx->uart, data_buf, max_len );
 }
@@ -187,8 +185,6 @@ gps_error_t gps_generic_parser
     
     return GPS_NO_ERROR;
 }
-
-
 
 // ----------------------------------------------- PRIVATE FUNCTION DEFINITIONS
